# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WEBCLERK 3.0 API",
    "version": "3.0.0",
    "description": "API for managing full ERP system with role-based access, email verification, and JWT authentication."
  },
  "paths": {
    "/api/auth/login/": {
      "post": {
        "operationId": "auth_login_create",
        "description": "Login and receive JWT pair plus basic user claims in the envelope data.",
        "tags": [
          "api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "BearerAuth": {
              "type": "http",
              "scheme": "bearer",
              "bearerFormat": "JWT"
            }
          },
          {}
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginEnvelope"
                },
                "examples": {
                  "LoginSuccess": {
                    "value": {
                      "status": "success",
                      "error": null,
                      "code": 200,
                      "message": "",
                      "data": {
                        "refresh": "...",
                        "access": "...",
                        "email": "1@1.com",
                        "role": "admin",
                        "name_first": "first_1",
                        "name_last": "last_1"
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginEnvelope": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "error": {
            "nullable": true
          },
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/LoginResponse"
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "status"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "refresh": {
            "type": "string"
          },
          "access": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string"
          },
          "name_first": {
            "type": "string"
          },
          "name_last": {
            "type": "string"
          }
        },
        "required": [
          "access",
          "email",
          "name_first",
          "name_last",
          "refresh",
          "role"
        ]
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      },
      "jwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}